backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
SQLAlchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
python-jose[cryptography]==3.3.0 